Addition of brightness cutoff controller from 0.0 to 0.1 for lower value
